+++ src/tcNode/Properties/AssemblyInfo.cs
@@ -1,10 +1,19 @@
-﻿using System.Reflection;
+﻿/**************************************************************************************
+Trade Control Node Configuration
+
+Date: 1.8.19
+Author: Ian Monnox
+
+Documentation on github.com/tradecontrol/tcnodecore
+
+Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
+
+**************************************************************************************/
+
+using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
 [assembly: AssemblyTitle("Trade Control Node Library")]
 [assembly: AssemblyDescription("Installs and upgrades database instances")]
 [assembly: AssemblyConfiguration("")]
@@ -14,23 +23,9 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
 
-// The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("e0b55ebe-7f5e-4703-b983-a43bb9e3b8d7")]
 
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("3.25.1.0")]
-[assembly: AssemblyFileVersion("3.25.1.0")]
+[assembly: AssemblyVersion("3.26.1.0")]
+[assembly: AssemblyFileVersion("3.26.1.0")]