+++ src/tcNode/Properties/Resources.Designer.cs
@@ -300,7 +300,7 @@ namespace TradeControl.Node.Properties {
         ///Sample upgrade script
         ///Release: 3.25.1
         ///
-        ///Date: 25 February 2020
+        ///Date: 29 February 2020
         ///Author: Ian Monnox
         ///
         ///Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
@@ -319,6 +319,34 @@ namespace TradeControl.Node.Properties {
         internal static string tc_upgrade_3_25_1 {
             get {
                 return ResourceManager.GetString("tc_upgrade_3_25_1", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to /**************************************************************************************
+        ///Trade Control
+        ///Sample upgrade script
+        ///Release: 3.26.1
+        ///
+        ///Date: 06 April 2020
+        ///Author: Ian Monnox
+        ///
+        ///Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
+        ///
+        ///You may obtain a copy of the License at
+        ///
+        ///	https://www.gnu.org/licenses/gpl-3.0.en.html
+        ///
+        ///Change log:
+        ///
+        ///	https://github.com/tradecontrol/tc-nodecore
+        ///
+        ///Instructions:
+        ///This script should be applied by the TC Node Configuration app. [rest of string was truncated]&quot;;.
+        /// </summary>
+        internal static string tc_upgrade_3_26_1 {
+            get {
+                return ResourceManager.GetString("tc_upgrade_3_26_1", resourceCulture);
             }
         }
     }